split preferencesfragment
